Checkpoint Commit restructuring API with new Honcho
