+++ bot/app.py
@@ -1,10 +1,11 @@
 import discord
 import sentry_sdk
 import os
+import asyncio
 
 from common import init
 from dotenv import load_dotenv
-    
+
 load_dotenv()
 
 rate = 0.2 if os.getenv("SENTRY_ENVIRONMENT") == "production" else 1.0
@@ -12,10 +13,14 @@ rate = 0.2 if os.getenv("SENTRY_ENVIRONMENT") == "production" else 1.0
 sentry_sdk.init(
     dsn=os.getenv("SENTRY_DSN_DISCORD"),
     traces_sample_rate=rate,
-    profiles_sample_rate=rate
+    profiles_sample_rate=rate,
 )
 
 CACHE, LOCK, _, (THOUGHT_CHANNEL, TOKEN) = init()
+
+THOUGHT_CHANNEL = os.environ["THOUGHT_CHANNEL_ID"]
+TOKEN = os.environ["BOT_TOKEN"]
+LOCK = asyncio.Lock()
 
 intents = discord.Intents.default()
 intents.messages = True
