+++ pyproject.toml
@@ -16,8 +16,9 @@ pexpect = "^4.8.0"
 fastapi = "^0.101.1"
 uvicorn = {extras = ["standard"], version = "^0.23.2"}
 sentry-sdk = {extras = ["fastapi"], version = "^1.31.0"}
-honcho-ai = {path = "../honcho/sdk", develop = true}
+honcho-ai = "0.0.8a1"
 langchain-openai = "^0.0.8"
+mirascope = "^0.13.1"
 
 [build-system]
 requires = ["poetry-core"]
